Add --output-find-links options
This options is very similar to the existing --output-index-url option.
However, in our case we have a webserver with only a few additional
packages, and we don't mirror all the packages that we need from pypi.
In this case, we need to add a --find-index options to the requirements
file, not a --index-url.